s3s: Fix error code for invalid x-amz-content-sha256 header
Change the error code from SignatureDoesNotMatch to
XAmzContentSHA256Mismatch when the x-amz-content-sha256 header
fails to parse or is invalid.

This improves S3 compatibility by returning the same error code that
AWS S3 returns in this scenario. The two error codes serve different
purposes:
- XAmzContentSHA256Mismatch: returned when the custom header
  x-amz-content-sha256 does not match the computed hash of the
  request payload
- SignatureDoesNotMatch: returned when the signature value in the
  HTTP request does not match the one calculated by the S3 service

The comment is also updated to clarify this distinction.

Signed-off-by: Kefu Chai <tchaikov@gmail.com>